migrate notifications controller
