--- packages/server/src/app/api/notification/notification.controller.spec.ts
@@ -1,18 +1,0 @@
-import { Test, TestingModule } from '@nestjs/testing';
-import { NotificationController } from './notification.controller';
-
-describe('Notification Controller', () => {
-    let controller: NotificationController;
-
-    beforeEach(async () => {
-        const module: TestingModule = await Test.createTestingModule({
-            controllers: [NotificationController],
-        }).compile();
-
-        controller = module.get<NotificationController>(NotificationController);
-    });
-
-    it('should be defined', () => {
-        expect(controller).toBeDefined();
-    });
-});
